+++ src/Epidemic/InhomogeneousBD.hs
@@ -5,47 +5,56 @@ module Epidemic.InhomogeneousBD
   , allEvents
   ) where
 
+import Control.Monad (liftM)
+import Data.Maybe (fromJust)
 import qualified Data.Vector as V
 import System.Random.MWC
 import System.Random.MWC.Distributions (bernoulli, exponential)
-
 import Epidemic
 import Epidemic.Utility
 
-data BDRates =
-  BDRates Rate Rate
+data InhomBDRates =
+  InhomBDRates (Timed Rate) Rate
 
-instance ModelParameters BDRates where
-  rNaught (BDRates birthRate deathRate) _ = birthRate / deathRate
-  eventRate (BDRates birthRate deathRate) _ = birthRate + deathRate
+instance ModelParameters InhomBDRates where
+  rNaught (InhomBDRates timedBirthRate deathRate) time =
+    let birthRate = cadlagValue timedBirthRate time
+     in liftM (/ deathRate) birthRate
+  eventRate (InhomBDRates timedBirthRate deathRate) time =
+    let birthRate = cadlagValue timedBirthRate time
+     in liftM (+ deathRate) birthRate
+  birthProb (InhomBDRates timedBirthRate deathRate) time =
+    liftM (\br -> br / (br + deathRate)) $ cadlagValue timedBirthRate time
 
-newtype BDPopulation =
-  BDPopulation People
+newtype InhomBDPop =
+  InhomBDPop People
   deriving (Show)
 
-instance Population BDPopulation where
+instance Population InhomBDPop where
   susceptiblePeople _ = Nothing
-  infectiousPeople (BDPopulation people) = Just people
+  infectiousPeople (InhomBDPop people) = Just people
   removedPeople _ = Nothing
-  isInfected (BDPopulation (People people)) = not $ V.null people
+  isInfected (InhomBDPop people) = not $ nullPeople people
 
 -- | Return a BD-process parameters object
-birthDeathRates :: Rate -- ^ birth rate
-                -> Rate -- ^ death rate
-                -> Maybe BDRates
-birthDeathRates birthRate deathRate
-  | birthRate >= 0 && deathRate >= 0 = Just $ BDRates birthRate deathRate
+inhomBirthDeathRates :: [(Time, Rate)] -- ^ birth rate
+                     -> Rate           -- ^ death rate
+                     -> Maybe InhomBDRates
+inhomBirthDeathRates tBrPairs deathRate
+  | all (\x -> 0 < snd x) tBrPairs && deathRate >= 0 = liftM (\tbr -> InhomBDRates tbr deathRate) $ asTimed tBrPairs
   | otherwise = Nothing
 
--- | Configuration of a birth-death simulation.
-configuration :: Time         -- ^ Duration of the simulation
-                 -> (Rate, Rate) -- ^ Birth and Death rates
-                 -> Maybe (SimulationConfiguration BDRates BDPopulation)
-configuration maxTime (birthRate, deathRate) =
+-- | Configuration of a inhomogeneous birth-death simulation.
+configuration :: Time                     -- ^ Duration of the simulation
+              -> ([(Time,Rate)], Rate) -- ^ Birth and Death rates
+              -> Maybe (SimulationConfiguration InhomBDRates InhomBDPop)
+configuration maxTime (tBrPairs, deathRate) =
   let (seedPerson, newId) = newPerson initialIdentifier
-      bdPop = BDPopulation (People $ V.singleton seedPerson)
-   in do maybeBDRates <- birthDeathRates birthRate deathRate
-         if maxTime > 0 then Just (SimulationConfiguration maybeBDRates bdPop newId maxTime) else Nothing
+      bdPop = InhomBDPop (People $ V.singleton seedPerson)
+   in do maybeIBDRates <- inhomBirthDeathRates tBrPairs deathRate
+         if maxTime > 0
+           then Just (SimulationConfiguration maybeIBDRates bdPop newId maxTime)
+           else Nothing
 
 -- | A random event and the state afterwards
 randomEvent ::
@@ -55,22 +64,28 @@ randomEvent ::
   -> Identifier   -- ^ current identifier
   -> GenIO        -- ^ PRNG
   -> IO (Time, Event, InhomBDPop, Identifier)
-randomEvent inhomRates@(InhomBDRates brts dr) currTime (InhomBDPop currPeople) currId gen = do
-  delay <- inhomExponential (fromIntegral $ numPeople currPeople) inhomRates gen
-  isBirth <- bernoulli (birthProb inhomRates (currTime + delay)) gen
-  (selectedPerson, unselectedPeople) <- randomPerson currPeople gen
-  return $
-    if isBirth
-      then let newTime = currTime + delay
-               (birthedPerson, newId) = newPerson currId
-               event = InfectionEvent newTime selectedPerson birthedPerson
-            in ( newTime
-               , event
-               , InhomBDPop (People $ addPerson birthedPerson currPeople)
-               , newId)
-      else let newTime = currTime + delay
-               event = RemovalEvent newTime selectedPerson
-            in (newTime, event, BDPopulation (People unselectedPeople), currId)
+randomEvent inhomRates@(InhomBDRates brts dr) currTime (InhomBDPop (people@(People peopleVec))) currId gen =
+  let popSize = fromIntegral $ numPeople people :: Double
+      stepFunction = [(t-currTime,popSize * fromJust (eventRate inhomRates t)) | (t,br) <- brts]
+   in do delay <- inhomExponential stepFunction gen
+         isBirth <- bernoulli (fromJust (birthProb inhomRates (currTime + delay))) gen
+         (selectedPerson, unselectedPeople) <- randomPerson peopleVec gen
+         return $
+           if isBirth
+             then let newTime = currTime + delay
+                      (birthedPerson, newId) = newPerson currId
+                      event =
+                        InfectionEvent newTime selectedPerson birthedPerson
+                   in ( newTime
+                      , event
+                      , InhomBDPop (addPerson birthedPerson people)
+                      , newId)
+             else let newTime = currTime + delay
+                      event = RemovalEvent newTime selectedPerson
+                   in ( newTime
+                      , event
+                      , InhomBDPop (People unselectedPeople)
+                      , currId)
 
 -- | The state of the simulation at the time of the last event prior to the
 -- stopping time.